fixed group to be more generic
