+++ src/components/basic/form/Checkbox.md
@@ -17,14 +17,14 @@ Html Snippet:
 ### Checkbox Group
 
 ```jsx
-import CheckboxGroup from './CheckboxGroup'
+import CheckedInputGroup from './CheckedInputGroup'
 ;
-<CheckboxGroup hint="Select all that apply." describedBy="favorite-colors-hint" heading="What are your favorite colors?">
+<CheckedInputGroup hint="Select all that apply." describedBy="favorite-colors-hint" heading="What are your favorite colors?">
 	<Checkbox id="color-blue" name="favorite-color" label="Blue" />
 	<Checkbox id="color-red" name="favorite-color" label="Red" />
 	<Checkbox id="color-yellow" name="favorite-color" label="Yellow" />
 	<Checkbox id="color-black" name="favorite-color" label="Black" disabled checked />
-</CheckboxGroup>
+</CheckedInputGroup>
 ```
 
 Html Snippet:
